Update pxt.json, main.blocks, main.ts

+++ main.ts
@@ -48,23 +48,27 @@ Planet = game.createSprite(2, 2)
 Ship = game.createSprite(0, 0)
 sound = 1
 basic.forever(function () {
-    Planet.move(1)
-    Planet.ifOnEdgeBounce()
-    basic.pause(100)
-})
-basic.forever(function () {
-    basic.pause(100 * randint(1, 5))
-    Asteroid = game.createSprite(randint(0, 4), 4)
-    Asteroid.turn(Direction.Right, randint(0, 360))
-    for (let index = 0; index < 15; index++) {
-        Asteroid.move(1)
-        if (Asteroid.isTouching(Ship)) {
-            game.addScore(randint(10, 50))
-            music.playTone(131, music.beat(BeatFraction.Whole))
-            Asteroid.turn(Direction.Right, 180)
-        }
-        Asteroid.ifOnEdgeBounce()
+    if (!(game.isGameOver())) {
+        Planet.move(1)
+        Planet.ifOnEdgeBounce()
         basic.pause(100)
     }
-    Asteroid.delete()
+})
+basic.forever(function () {
+    if (!(game.isGameOver())) {
+        basic.pause(100 * randint(1, 5))
+        Asteroid = game.createSprite(randint(0, 4), 4)
+        Asteroid.turn(Direction.Right, randint(0, 360))
+        for (let index = 0; index < 15; index++) {
+            Asteroid.move(1)
+            if (Asteroid.isTouching(Ship)) {
+                game.addScore(randint(10, 50))
+                music.playTone(131, music.beat(BeatFraction.Whole))
+                Asteroid.turn(Direction.Right, 180)
+            }
+            Asteroid.ifOnEdgeBounce()
+            basic.pause(100)
+        }
+        Asteroid.delete()
+    }
 })
